+++ src/components/index.js
@@ -1,0 +1,11 @@
+export { PrimaryButton, SecondaryButton } from "./Buttons/Buttons";
+export { NewPost } from "./NewPost/NewPost";
+export { PostCard } from "./PostCard/PostCard";
+export { PostOptionsModal } from "./PostOptionsModal/PostOptionsModal";
+export { SearchBar } from "./SearchBar/SearchBar";
+export { SearchedUsersModal } from "./SearchedUsersModal/SearchedUsersModal";
+export { SideBar } from "./SideBar/SideBar";
+export { SortBar } from "./SortBar/SortBar";
+export { SuggestedUsers } from "./SuggestedUsers/SuggestedUsers";
+export { ToastWrapper } from "./ToastWrapper/ToastWrapper";
+export { UserAvatar } from "./UserAvatar/UserAvatar";
