+++ src/utils/icons.js
@@ -1,0 +1,47 @@
+import {
+  FaImage,
+  FaSmile,
+  FaRegHeart,
+  FaHeart,
+  FaRegComments,
+  FaRegBookmark,
+  FaFilter,
+  FaFire,
+  FaArrowUp,
+  FaArrowDown,
+} from "react-icons/fa";
+import {
+  MdCancel,
+  MdShare,
+  MdHome,
+  MdExplore,
+  MdBookmark,
+  MdPerson,
+} from "react-icons/md";
+import { HiDotsHorizontal, HiPlusCircle } from "react-icons/hi";
+import { FiSearch } from "react-icons/fi";
+import { AiOutlineEye, AiOutlineEyeInvisible } from "react-icons/ai";
+
+export {
+  FaImage,
+  FaSmile,
+  FaRegHeart,
+  FaHeart,
+  FaRegComments,
+  FaRegBookmark,
+  FaFilter,
+  FaFire,
+  FaArrowUp,
+  FaArrowDown,
+  FiSearch,
+  MdCancel,
+  MdShare,
+  MdHome,
+  MdExplore,
+  MdBookmark,
+  MdPerson,
+  HiDotsHorizontal,
+  HiPlusCircle,
+  AiOutlineEye,
+  AiOutlineEyeInvisible,
+};
